% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{clear}
\alias{clear}
\title{Clear (empty) a container}
\usage{
clear(container)
}
\arguments{
\item{container}{A container object.}
}
\description{
See \code{\link{empty}()} to check whether a container is empty.
}
