% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Counter object inspired on Python collections.Counter}
\usage{
counter(...)
}
\arguments{
\item{...}{Initial values of the counter.}
}
\description{
Counter object inspired on Python collections.Counter
}
