% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazySequence.R
\name{cycleSequence}
\alias{cycleSequence}
\title{Cycling sequence of integers}
\usage{
cycleSequence(start, end, step = 1L)
}
\arguments{
\item{start}{Start of the cycle, inclusive.}

\item{end}{End of the cycle, not included.}

\item{step}{The step size, defaults to \code{1L}.}
}
\value{
Cycle sequence object of integers cycling from start to end by step
size.
}
\description{
Cycling sequence of integers
}
\examples{
seq <- cycleSequence(1, 4, 2)
for (i in 1:4) {
  print(nextItem(seq))
}
}
