% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Dictionary object}
\usage{
dictionary()
}
\description{
A dictionary object backed by C++'s \code{std::map}. This is implemented
using a red black tree in most cases.
}

