% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{empty}
\alias{empty}
\title{Check whether the given container is empty}
\usage{
empty(container)
}
\arguments{
\item{container}{A container object.}
}
\description{
See \code{\link{clear}()} to clear a container.
}
