% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{filter}
\alias{filter}
\title{Filter a container using \code{fun()}}
\usage{
filter(container, fun)
}
\arguments{
\item{container}{The container to filter.}

\item{fun}{A function that returns a logical based on a value in the
container.}
}
\value{
The container after filtering.
}
\description{
Filter a container using \code{fun()}
}
