% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{getIndex}
\alias{getIndex}
\title{Get the current index of a sequence}
\usage{
getIndex(sequence)
}
\arguments{
\item{container}{A lazy sequence.}
}
\value{
The index.
}
\description{
Get the current index of a sequence
}
