% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{getValue}
\alias{getValue}
\title{Get a value from a container}
\usage{
getValue(container, key)
}
\arguments{
\item{container}{A container object.}

\item{key}{The key to get.}
}
\value{
The value.
}
\description{
Get a value from a container
}
