% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazySequence.R
\name{integerSequence}
\alias{integerSequence}
\title{Lazy sequence of natural numbers}
\usage{
integerSequence()
}
\value{
A lazy sequence of natural numbers.
}
\description{
Lazy sequence of natural numbers
}
\examples{
seq <- integerSequence()
for (i in 1:4) {
  print(nextItem(seq))
}
}
