% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{map}
\alias{map}
\title{Map the values in the container to new values suing \code{fun()}}
\usage{
map(container, fun)
}
\arguments{
\item{container}{The container to map.}

\item{fun}{A function that returns a new value for each value in the
container.}
}
\value{
The container with the values transformed using map.
}
\description{
Map the values in the container to new values suing \code{fun()}
}
