% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{nextItem}
\alias{nextItem}
\title{Get the next value in a sequence}
\usage{
nextItem(sequence)
}
\arguments{
\item{sequence}{A lazy sequence.}
}
\value{
The next item.
}
\description{
Get the next value in a sequence
}
