% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{peek}
\alias{peek}
\title{Peek at the next value in a sequence but don't advance in the sequence}
\usage{
peek(sequence)
}
\arguments{
\item{sequence}{A lazy sequence.}
}
\value{
The next item.
}
\description{
Peek at the next value in a sequence but don't advance in the sequence
}
