% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazySequence.R
\name{primeSequence}
\alias{primeSequence}
\title{Lazy sequence of primes}
\usage{
primeSequence()
}
\value{
Lazy sequence of primes.
}
\description{
Lazy sequence of primes
}
\examples{
seq <- primeSequence()
for (i in 1:4) {
  print(nextItem(seq))
}
}
