% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorityQueue.R
\name{priorityQueue}
\alias{priorityQueue}
\title{Priority Queue object}
\usage{
priorityQueue(...)
}
\arguments{
\item{...}{Initial values of the priority queue.}
}
\description{
A priority queue backed by C++'s \code{std::priority_queue}.
}
