% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{push}
\alias{push}
\title{Push a value to a container}
\usage{
push(container, value)
}
\arguments{
\item{container}{A container object.}

\item{value}{Value to add to the container.}
}
\description{
Push a value to a container
}
