% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{setValue}
\alias{setValue}
\title{Set a value with the given key in the given container}
\usage{
setValue(container, key, value)
}
\arguments{
\item{container}{A container object.}

\item{key}{The key to set.}

\item{value}{The value to set.}
}
\description{
Set a value with the given key in the given container
}
