% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazySequence.R
\name{stepSequence}
\alias{stepSequence}
\title{Lazy sequence object of evenly spaced integers}
\usage{
stepSequence(start, step = 1L)
}
\arguments{
\item{start}{The start of the sequence.}

\item{step}{The step size of the sequence, defaults to \code{1L}.}
}
\value{
Lazy sequence object of evenly spaced integers starting at
\code{start} increasing by \code{step}..
}
\description{
Lazy sequence object of evenly spaced integers
}
\examples{
seq <- stepSequence(1, 4)
for (i in 1:4) {
  print(nextItem(seq))
}
}
