% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringBuilder.R
\name{stringBuilder}
\alias{stringBuilder}
\title{String Builder object}
\usage{
stringBuilder(...)
}
\arguments{
\item{...}{Initial values of the priority queue.}
}
\description{
A StringBuilder backed by C++'s \code{std::vector<std::string>} and
\code{std::stringstream}.
}
