% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{take}
\alias{take}
\title{Take n elements}
\usage{
take(sequence, n)
}
\arguments{
\item{sequence}{A lazy sequence.}

\item{n}{The number of elements to take.}
}
\description{
Take n elements
}
\examples{
seq <- primeSequence()
unlist(take(seq, 10))
}
