% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{takeWhile}
\alias{takeWhile}
\title{Take elements while the condition is true}
\usage{
takeWhile(sequence, condition)
}
\arguments{
\item{sequence}{A lazy sequence.}

\item{condition}{Take numbers until this condition function is \code{FALSE}.}
}
\description{
Take elements while the condition is true
}
\examples{
seq <- primeSequence()
unlist(takeWhile(seq, \(x) x < 100L))
}
