% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{toList.Counter}
\alias{toList.Counter}
\title{Convert a Counter to a list}
\usage{
\method{toList}{Counter}(container)
}
\arguments{
\item{container}{The Counter to convert.}
}
\value{
A list.
}
\description{
Convert a Counter to a list
}
